requirements.txt
Django==1.8.18
django-model-utils
psycopg2-binary==2.7.4
pyaml==17.12.1


# Legacy main app requirements
timeside==0.7 # See https://github.com/Parisson/TimeSide/issues/104
git+https://github.com/Parisson/Telemeta.git@9c6a77974b338781db2971d66c96834b94ab2bb8
django-google-tools==1.1.0
MySQL-python==1.2.5
git+https://github.com/mariocesar/sorl-thumbnail.git@7ce76bc1ef798a63cbf89b5df83de4da4b12e98d#egg=sorl-thumbnail-master
git+https://github.com/Parisson/ebooklib.git#egg=ebooklib-0.16
git+https://github.com/Parisson/saved_searches.git@dj1.8#egg=saved_searches-2.0.0-beta
git+https://github.com/django-haystack/django-haystack.git@v2.5.1#egg=django-haystack-2.5.1
git+https://github.com/yomguy/django-ipauth.git#egg=django-ipauth-0.4.1-dev
git+https://github.com/Parisson/django-jqchat.git@dj1.8#egg=django-jqchat
django-registration-redux
django-extensions
django-timezones
django-debug-toolbar==1.6
django-extra-views
django-breadcrumbs
django-bootstrap3==8.1.0
django-bootstrap-pagination
django-json-rpc
django-suit
django-celery
docutils
psutil
pyyaml
python-ebml
zipstream
elasticsearch==1.6.0
redis
Werkzeug
django-dirtyfields==1.2.1
sqlparse==0.2.2



# Docker compose dependeny
docker-compose==1.19.0


# API dependencies
django-environ==0.4.4
djangorestframework==3.6.4
django-rest-swagger==2.1.2


# Testing dependencies
gunicorn==19.7.1
factory_boy==2.10.0
ipython==5.5.0
parameterized==0.6.1
pytest==3.4.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-env==0.6.2
pytest-pythonpath==0.7.2
